### 2017-11-20 #443 @Kevin-Mattheus-Moerman
@arfon if there are no references, do we need an empty paper.bib file?
