### 2017-11-20 #443 @Kevin-Mattheus-Moerman
@pzwang @rhiever @tacaswell @wesm @jreback @bryevdv is this [JOSS](http://joss.theoj.org/) submission your cup of tea? Would you be willing to review this?
