### 2017-11-20 #443 @arfon
> @arfon if there are no references, do we need an empty paper.bib file?

I guess not but in this case the `paper.md` file should not be pointing to a `paper.bib` file that doesn't exist. IMO the current paper is rather too brief and lacking context. I'd encourage the author to flesh this out (and perhaps cite related work/packages?)
