### 2017-11-20 #443 @ResidentMario
I'm happy to edit the write-up for content, I wasn't too sure how much is enough or too much. In terms of citations, there's an R package that was inspired by this Python package [here](https://www.rdocumentation.org/packages/poppr/versions/2.5.0/topics/missingno), as well as [this blog post](http://www.residentmar.io/2016/06/12/null-and-missing-data-python.html) describing the issue the package tackles more generally. Hopefully that can cover references.
