### 2017-11-26 e2cc2a0
Correct DatetimeIndex call.
