### 2018-01-26 #443 @Kevin-Mattheus-Moerman
@zkamvar would you be interested in reviewing this submission for JOSS?
