### 2018-01-26 #443 @Kevin-Mattheus-Moerman
@rhiever thanks for your reply. We are currently in a pre-review issue. If you agree to review I'll open a review issue and you'll be able to see the paper as well (the author here still needs to update their paper).

Here is an example of a [review issue](https://github.com/openjournals/joss-reviews/issues/500#event-1442337848) with 3 reviewers taking part. Essentially reviewers tick boxes at the top of the review issue. From the tickboxes you can see most of the review focuses on the software functionality and documentation.

Papers are typically very short (e.g. see [the paper](https://github.com/openjournals/joss-papers/blob/joss.00500/joss.00500/10.21105.joss.00500.pdf) submitted to the review issue I linked to).

You can also read more about our [reviewer guidelines](http://joss.theoj.org/about#reviewer_guidelines).
