### 2018-01-26 #443 @Kevin-Mattheus-Moerman
@ResidentMario please expand [your paper](
https://github.com/ResidentMario/missingno/blob/master/paper.md) as per [our guidelines](http://joss.theoj.org/about#author_guidelines). You can also check out [some accepted papers](http://joss.theoj.org/papers/accepted) to get a better picture of the content and level of detail presented in them.
