### 2018-01-26 #443 @ResidentMario
@Kevin-Mattheus-Moerman :+1: To review, I've gotten a smattering inquiries on how to cite this library, and this seemed like the perfect venue to achieve that. I made a pretty minimal submission to test the waters. V. happy to flush it out more.
