### 2018-01-26 #443 @zkamvar
Hello! Yes, I would be happy to review this.

Caveats:

1. I am primarily an R developer and use python on more of a monthly basis.
2. I am the author of the R package that @ResidentMario mentioned in [this comment](https://github.com/openjournals/joss-reviews/issues/443#issuecomment-345762333) (Though I don't see a citation necessary as the name was derived from [a pokemon glitch](https://en.wikipedia.org/wiki/MissingNo.)).
