### 2018-01-26 #547 @Kevin-Mattheus-Moerman
@rhiever and @zkamvar thank you for acting as reviewers! 🚀

There are two sets of tick boxes at the top of this issue, one for each of you. If you have minor comments to the author feel free to make them here. For bigger issues you may want to open an issue on this projects [repository](https://github.com/ResidentMario/missingno) which you can refer to here in the review issue.
Let me know if you have any questions. Also let me know if you are unable to tick the boxes.
