### 2018-01-26 ResidentMario/missingno#43 @Kevin-Mattheus-Moerman
Expand paper to conform to JOSS standards
See guidelines here:
http://joss.theoj.org/about#author_guidelines

See example accepted papers here:
http://joss.theoj.org/papers/accepted
