### 2018-01-26 #547 @Kevin-Mattheus-Moerman
@ResidentMario I've opened an issue about expanding your paper [here](https://github.com/ResidentMario/missingno/issues/43). Let me know if you have questions.
