+++ paper.bib
@@ -1,0 +1,6 @@
+@misc{missingno_archive,
+author = {{missingno Archive}},
+title = {Missingno: a missing data visualization suite},
+doi = {},
+howpublished = {\url{}}
+}
