+++ paper.bib
@@ -1,6 +1,7 @@
-@misc{missingno_archive,
-author = {{missingno Archive}},
-title = {Missingno: a missing data visualization suite},
-doi = {},
-howpublished = {\url{}}
+@online{null-data-blog-post,
+  author = {Aleksey Bilogur},
+  title = {Null and missing data in Python},
+  year = 2016,
+  url = {http://www.residentmar.io/2016/06/12/null-and-missing-data-python.html},
+  urldate = {2016-06-12}
 }
