### 2018-01-29 ResidentMario/missingno#43 @ResidentMario
Paper expanded. I'll leave this issue open for now, though.
