### 2018-01-29 #547 @ResidentMario
@Kevin-Mattheus-Moerman Quick question: the review instructions cite version 0.3.7, but the package is currently at version 0.3.8. What's the fix?

@zkamvar Can you link me to the R package you cited [here](https://github.com/openjournals/joss-reviews/issues/443#issuecomment-360814979)? I've lost my reference to it and want to add it to the references per [this comment](https://github.com/openjournals/joss-reviews/issues/443#issuecomment-345572443).
