### 2018-01-29 #547 @zkamvar
Hi @ResidentMario, The citation for the package (which includes the reference for the missingno function) is here: https://peerj.com/articles/281/, though I think it would be better to [cite the SciPy packages](https://scipy.org/citing.html) you import (I'll open an issue for this).
