### 2018-01-29 ResidentMario/missingno#43 @Kevin-Mattheus-Moerman
Thanks, looks a lot better. Here are some quick points on the paper.
- [ ] Please work on adding references
- [ ] Fix type: "..easily to, at a glance, understanding patterns..." should probably be: "...easily to, at a glance, understand patterns'''"
- [ ] Fix typo: "provides an methodology" should be : "provides a methodology"
- [ ] To me something feels odd/missing here due to the use of the word both: "The visualizations are consciously designed to be as effective as possible at uncovering missing data patterns both between columns of data, and hence, to help its users build more effective data models and pipelines."
- [ ] I suggest and recommend (since your package offers visualization), to add a figure which you could use as a "graphical summary" of what your package does.
