### 2018-01-29 #547 @Kevin-Mattheus-Moerman
@ResidentMario @zkamvar @rhiever
[📖 The paper](https://github.com/openjournals/joss-papers/blob/joss.00547/joss.00547/10.21105.joss.00547.pdf) now renders as a PDF. I've added starting commenting and providing recommendations on the paper in [this issue](https://github.com/ResidentMario/missingno/issues/43). Feel free to join there in relation to purely the paper.
