### 2018-01-29 ResidentMario/missingno#44 @zkamvar
Note: this is related to the first item in https://github.com/ResidentMario/missingno/issues/43#issuecomment-361284523
