### 2018-01-29 #547 @zkamvar
I have opened an issue (https://github.com/ResidentMario/missingno/issues/45) regarding the inclusion of example data internal to the package.
