### 2018-01-29 ResidentMario/missingno#45 @ResidentMario
I had success using the Quilt package manager (https://quiltdata.com/) for this task for my `geoplot` project. Reckon I'll try to do that again with this dataset. Been meaning to do it for a while, but haven't gotten to it before.

Having the complete dataset is important in the last demo in particular: the geospatial demo really does require all that data to show a "dense enough" outline of the city.

I'll wrangle something up.
