### 2018-01-29 ResidentMario/missingno#46 @zkamvar
Add tests for nullity_filter and nullity_sort
Since these functions return a pandas data frame, they can be tested for correctness. Please include a testing suite for these functions.
