### 2018-01-29 #547 @zkamvar
I have opened an issue (https://github.com/ResidentMario/missingno/issues/46) regarding tests on non-visualization functions.
