### 2018-01-29 ResidentMario/missingno#45 @zkamvar
To clarify a bit:

I'm not suggesting you get rid of the large examples all together. I think it would be beneficial to have a small example for the quick start and then move on to the larger example to show the real power of the package.
