### 2018-01-29 #547 @Kevin-Mattheus-Moerman
@arfon the author (@ResidentMario) asked

> Quick question: the review instructions cite version 0.3.7, but the package is currently at version 0.3.8. What's the fix?

Can you help update the version listed or can the user or I do this ourselves?
