### 2018-01-29 ResidentMario/missingno#46 @zkamvar
N.B. This also applies to helper functions such as [`_calculate_geographic_nullity()`](https://github.com/ResidentMario/missingno/blob/4230493677207ae8474f6e8f6a7adec4b528036a/missingno/missingno.py#L573-L589)
