### 2018-01-29 #547 @zkamvar
Question for @Kevin-Mattheus-Moerman: When I'm finished with the review, would you like me to give a brief summary of my assessment (e.g.: https://github.com/openjournals/joss-reviews/issues/514#issuecomment-357365641)?
