### 2018-01-29 #547 @zkamvar
@ResidentMario, this is a fantastic package. It is straightforward enough that I was able to use it for my own data without much effort 😃👍. The documentation is clear and the examples are wonderful for showing the utility of the package. Because of a lack of tests (though this is partially explained in the CONTRIBUTING.md file), and citations to relevant python software, I would recommend this for acceptance with **minor revisions**.

### Good points

 - easy to install with pip
 - built on top of existing, trusted packages
 - easy to use
 - clearly documented

### Blockers

 - [ ] no tests for testable functions (https://github.com/ResidentMario/missingno/issues/46)
 - [ ] paper needs citations to relevant packages (https://github.com/ResidentMario/missingno/issues/44)

### Good to have

 - [ ] a small example data set packaged with missingno  for quick demonstration and testing. (https://github.com/ResidentMario/missingno/issues/45)

<summary>My Python setup (3.4.4)</summary>

```
Python 3.4.4 |Anaconda custom (x86_64)| (default, Jan  9 2016, 17:30:09)
[GCC 4.2.1 (Apple Inc. build 5577)] on darwin
```
