### 2018-01-30 d856eca
Remove stray comments, add a utils test suite.
