### 2018-01-31 ResidentMario/missingno#45 @rhiever
Big 👍 to this suggestion from @zkamvar. It took far too long for me to download the example dataset (~250 MB for the one I downloaded). A small dataset (<1 MB) crafted for demo purposes would be ideal, such that users can install your package and instantly run quickstart examples from the README. Very important for uptake of projects.
