### 2018-01-31 ResidentMario/missingno#48 @rhiever
Create a development branch
I strongly suggest that you create a development branch for this repo. That way you can develop the next release on the development branch and maintain the master branch for stable releases.
