### 2018-01-31 ResidentMario/missingno#49 @rhiever
Add badges to your README to show supported Python versions + software license
Badges are useful for showing the status of your project and related information in a quick, visual manner. I think it would be beneficial to list information such as the versions of Python supported by your project (2.7? 3.6?) and software license (MIT).
