### 2018-01-31 ResidentMario/missingno#50 @rhiever
Add grid lines to the heatmap for interpretability
Without grid lines, I have difficult interpreting the heatmap. It would be helpful from an interpretability perspective to add gridlines that we can follow when matching columns and rows.

![image](https://user-images.githubusercontent.com/1719223/35602883-7a242e68-05ef-11e8-842e-1907735b6476.png)
