### 2018-01-31 ResidentMario/missingno#51 @rhiever
Could not reproduce heatmap from the README
I download and processed the collisions dataset using the notebook you link to in the README. I then fed that processed collisions dataset to the heatmap function (missingno v0.3.8) and this was my output:

![heatmap-error](https://user-images.githubusercontent.com/1719223/35602908-9d8ded6c-05ef-11e8-85ec-7f403e5f5959.png)

For some reason, the grid cells in the heatmap that don't have "significant" values aren't being masked. Happy to provide other package versions if that would be useful for debugging.
