### 2018-01-31 ResidentMario/missingno#52 @rhiever
shapely and descartes should be dependencies?
Shouldn't `shapely` and `descartes` be dependencies? I need them if I want to use the geographical plotting capabilities of the library, which sounds like it makes them package dependencies to me.
