### 2018-01-31 ResidentMario/missingno#53 @rhiever
Suggestion: inline=False by default for plotting functions
I suggest that the default for `inline` should be `False` for the plotting functions. I commonly assume that the plot I just generated via any Python dataviz function can be manipulated via `matplotlib.pyplot`, or at least the function will return the figure to manipulate further. I assume that many users would think that way too, given the behavior of matplotlib (of course), Seaborn, etc.

The most common use case I can imagine is to save the figure, which AFAICT can't be done with missingno without setting `inline=False`.
