### 2018-01-31 ResidentMario/missingno#54 @rhiever
README is too crowded - reorganize into a separate GitHub pages docs site
I suggest keeping the README short and to the point: badges showing status etc. of the package, what is the purpose of the package, how to install, 1-2 basic code examples, and contributing information.

More examples, function API docs, and detailed descriptions can be moved to the docs site.

I also suggest adding a gallery to the docs, [similar to Seaborn](https://seaborn.pydata.org/examples/index.html).
