### 2018-01-31 ResidentMario/missingno#55 @rhiever
Suggestion: Move __version__ variable to a separate file
I suggest moving the `__version__` variable to a separate `_version.py` file so that the variable doesn't get lost in the rest of the core functionality of the package. No need to store packaged-related information in the main code file(s).
