### 2018-01-31 ResidentMario/missingno#56 @rhiever
Suggestion: Break up helper functions and plotting functions into separate files
`missingno.py` is getting pretty long in terms of SLOC, and the easiest way to divide the package into meaningful separate files would be to break the helper functions out into a `helper.py` file. There's no need for these functions to rest in the same file.
