### 2018-01-31 ResidentMario/missingno#57 @rhiever
Mix of ' and " quotes in the code
There is a mix of ' and " quotes in the code. For code quality purposes, choose one and stick with it throughout the package. I recommend '.
