### 2018-01-31 ResidentMario/missingno#58 @rhiever
Suggestion: Remove obvious / extraneous code comments
I suggest you remove obvious comments such as "# Display." and "# Now we draw.", or make them more useful. Comments should be reserved for non-obvious software logic.
