### 2018-01-31 ResidentMario/missingno#59 @rhiever
geoplot function would not work for me
I installed the optional dependencies then used the same data from #51. I passed it to the `geoplot` function examples from the README and these are the outputs.

![geoplot-error1](https://user-images.githubusercontent.com/1719223/35603378-eeb5bc4a-05f1-11e8-8574-591ad89e9c04.png)

![geoplot-error2](https://user-images.githubusercontent.com/1719223/35603379-f1dc31e2-05f1-11e8-8457-93f63fb40bef.png)

I don't do much geospatial dataviz in Python, so it's possible I'm missing another dependency or have a bad setting. Please let me know what package versions you would like to know, if that information is useful.
