### 2018-01-31 ResidentMario/missingno#60 @rhiever
geoplot function imports packages every time it is called
The `geoplot` function imports a handful of packages every time it's called. I don't recommend doing this. I suggest you place all imports at the top of the code file.

Alternatively, place the `geoplot` function in a separate file/subdirectory of the package with the imports at the top such that the imports are invoked only when `geoplot` is imported.

This issue is somewhat related to #52.
