### 2018-01-31 ResidentMario/missingno#61 @rhiever
Add continuous integration
Once #46 and #47 are resolved, it will be invaluable for you to set up continuous integration via [Travis-CI](https://travis-ci.org) or related services. That way every commit, PR, etc. will automatically be tested for you.

If you're unsure how to add CI to a GitHub repo, see the [GitHub docpage](https://github.com/marketplace/category/continuous-integration) or ping me on this issue. These services are free for public repos.
