### 2018-01-31 ResidentMario/missingno#62 @rhiever
Suggest adding test coverage and code quality trackers
[landscape.io](https://landscape.io) is helpful for tracking test coverage for your package.

[coveralls.io](https://coveralls.io) is helpful for linting and tracking the code quality for your package.

Both are free services for public repos and can be paired with the continuous integration services mentioned in #61.
