### 2018-01-31 #547 @rhiever
Thank you for the great package, @ResidentMario. I was able to run and verify most of the examples from the README and can clearly see the utility of this package for my own work. Based on my evaluation of the package, I recommend several **minor revisions** as listed below.

### Good points

* easy installation via pip
* clean, useful plots right out of the box
* good documentation and examples
* minimal dependencies, and the dependencies are commonly-used packages

### Blockers

Agree with @zkamvar's blockers. Adding my own:

- [ ] Create a development branch on the repo (https://github.com/ResidentMario/missingno/issues/48)
- [ ] Add grid lines to the heatmap for interpretability (https://github.com/ResidentMario/missingno/issues/50)
- [ ] Fix / address heatmap reproducibility bug (https://github.com/ResidentMario/missingno/issues/51)
- [ ] Fix / address issues with `geoplot` function not working (https://github.com/ResidentMario/missingno/issues/59)
- [ ] Address `geoplot` function imports (https://github.com/ResidentMario/missingno/issues/60)

### Good to have

Strong 👍 on @zkamvar's suggestion of having a small example dataset for the examples. I might even elevate that issue to a Blocker.

- [ ] Unit tests for dataviz functions (idea: https://github.com/ResidentMario/missingno/issues/47)
- [ ] Add more useful badges to the repo (https://github.com/ResidentMario/missingno/issues/49)
- [ ] Sort out whether `shapely` and `descartes` are actually optional dependencies (https://github.com/ResidentMario/missingno/issues/52)
- [ ] Change `inline` parameter default to `False` (https://github.com/ResidentMario/missingno/issues/53)
- [ ] Shorten and refine the README and create separate docs pages (https://github.com/ResidentMario/missingno/issues/54)
- [ ] Move `__version__` variable out of the main source file (https://github.com/ResidentMario/missingno/issues/55)
- [ ] Break up helper functions and plotting functions into separate files (https://github.com/ResidentMario/missingno/issues/56)
- [ ] Decide on usage of ' and " (https://github.com/ResidentMario/missingno/issues/57)
- [ ] Clean up code comments (https://github.com/ResidentMario/missingno/issues/58)
- [ ] Add continuous integration + testing (https://github.com/ResidentMario/missingno/issues/61)
- [ ] Add test coverage and code quality trackers (https://github.com/ResidentMario/missingno/issues/62)

<summary>My Python setup (3.6.4):</summary>

```
Python 3.6.4 |Anaconda custom (64-bit)| (default, Jan 16 2018, 12:04:33)
Type 'copyright', 'credits' or 'license' for more information
IPython 6.2.1 -- An enhanced Interactive Python. Type '?' for help.
```
