### 2018-02-02 0787a81
Clean up helpers, remove stray comments, add some material.
