+++ paper.bib
@@ -1,7 +1,41 @@
 @online{null-data-blog-post,
-  author = {Aleksey Bilogur},
-  title = {Null and missing data in Python},
-  year = 2016,
-  url = {http://www.residentmar.io/2016/06/12/null-and-missing-data-python.html},
-  urldate = {2016-06-12}
+    author = {Aleksey Bilogur},
+    title = {Null and missing data in Python},
+    year = 2016,
+    url = {http://www.residentmar.io/2016/06/12/null-and-missing-data-python.html},
+    urldate = {2016-06-12}
+}
+
+@Misc{scipy,
+    author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
+    title =     {{SciPy}: Open source scientific tools for {Python}},
+    year =      {2001--},
+    url = "http://www.scipy.org/",
+    note = {[Online; accessed <today>]}
+}
+
+@article{numpy,
+    author={Stéfan van der Walt, S. Chris Colbert and Gaël Varoquaux},
+    title={The NumPy Array: A Structure for Efficient Numerical Computation},
+    journal={Computing in Science & Engineering},
+    volume=13,
+    pages=22-30,
+    year=2011
+}
+
+@article{matplotlib,
+    author={John D. Hunter},
+    title={Matplotlib: A 2D Graphics Environment},
+    year=2007,
+    volume=9,
+    pages=90-95,
+    journal={Computing in Science & Engineering}
+}
+
+@article{pandas,
+    author={Wes McKinney},
+    title={Data Structures for Statistical Computing in Python},
+    journal={Proceedings of the 9th Python in Science Conference},
+    year=2010,
+    pages=51-56
 }
