### 2018-02-02 2e3a4e4
Clean up comments, factor out __version__.
