+++ missingno/__init__.py
@@ -5,4 +5,4 @@ from .missingno import dendrogram
 from .missingno import geoplot
 from .missingno import nullity_filter
 from .missingno import nullity_sort
-from .missingno import __version__
+from ._version import __version__
