+++ missingno/missingno.py
@@ -6,9 +6,6 @@ from scipy.cluster import hierarchy
 import seaborn as sns
 import pandas as pd
 from mpl_toolkits.axes_grid1 import make_axes_locatable
-import warnings
-
-__version__ = "0.3.8"
 
 
 def nullity_sort(df, sort=None):
@@ -19,9 +16,9 @@ def nullity_sort(df, sort=None):
     :param sort: The sorting method: either "ascending", "descending", or None (default).
     :return: The nullity-sorted DataFrame.
     """
-    if sort == "ascending":
+    if sort == '"ascending':
         return df.iloc[np.argsort(df.count(axis='columns').values), :]
-    elif sort == "descending":
+    elif sort == 'descending':
         return df.iloc[np.flipud(np.argsort(df.count(axis='columns').values)), :]
     else:
         return df
@@ -43,12 +40,12 @@ def nullity_filter(df, filter=None, p=0, n=0):
     :param n: A numerical cut-off. If non-zero no more than this number of columns will be returned.
     :return: The nullity-filtered `DataFrame`.
     """
-    if filter == "top":
+    if filter == 'top':
         if p:
             df = df.iloc[:, [c >= p for c in df.count(axis='rows').values / len(df)]]
         if n:
             df = df.iloc[:, np.sort(np.argsort(df.count(axis='rows').values)[-n:])]
-    elif filter == "bottom":
+    elif filter == 'bottom':
         if p:
             df = df.iloc[:, [c <= p for c in df.count(axis='rows').values / len(df)]]
         if n:
@@ -94,27 +91,20 @@ def matrix(df,
     :param color: The color of the filled columns. Default is a medium dark gray: the RGB multiple `(0.25, 0.25, 0.25)`.
     :return: If `inline` is True, the underlying `matplotlib.figure` object. Else, nothing.
     """
-
-    # Apply filters and sorts.
     df = nullity_filter(df, filter=filter, n=n, p=p)
     df = nullity_sort(df, sort=sort)
 
     height = df.shape[0]
     width = df.shape[1]
 
-    # z is the color-mask array.
+    # z is the color-mask array, g is a NxNx3 matrix. Apply the z color-mask to set the RGB of each pixel.
     z = df.notnull().values
-
-    # g is a NxNx3 matrix
     g = np.zeros((height, width, 3))
 
-    # Apply the z color-mask to set the RGB of each pixel.
     g[z < 0.5] = [1, 1, 1]
     g[z > 0.5] = color
 
-    # Set up the matplotlib grid layout.
-    # If the sparkline is removed the layout is a unary subplot.
-    # If the sparkline is included the layout is a left-right subplot.
+    # Set up the matplotlib grid layout. A unary subplot if no sparkline, a left-right splot if yes sparkline.
     fig = plt.figure(figsize=figsize)
     if sparkline:
         gs = gridspec.GridSpec(1, 2, width_ratios=width_ratios)
@@ -138,9 +128,8 @@ def matrix(df,
     ax0.spines['bottom'].set_visible(False)
     ax0.spines['left'].set_visible(False)
 
-    # Set up and rotate the column ticks.
-    # The labels argument is set to None by default. If the user specifies it in the argument,
-    # respect that specification. Otherwise display for <= 50 columns and do not display for > 50.
+    # Set up and rotate the column ticks. The labels argument is set to None by default. If the user specifies it in
+    # the argument, respect that specification. Otherwise display for <= 50 columns and do not display for > 50.
     if labels or (labels is None and len(df.columns) <= 50):
         ha = 'left'
         ax0.set_xticks(list(range(0, width)))
@@ -148,8 +137,7 @@ def matrix(df,
     else:
[…]
