### 2018-02-02 ResidentMario/missingno#58 @ResidentMario
Done in 2e3a4e4.
