### 2018-02-02 2f8737a
Separate out utility functions.
