+++ missingno/__init__.py
@@ -6,3 +6,4 @@ from .missingno import geoplot
 from .missingno import nullity_filter
 from .missingno import nullity_sort
 from ._version import __version__
+from .utils import nullity_filter, nullity_sort
