### 2018-02-02 ResidentMario/missingno#56 @ResidentMario
Done in 2f8737a (what-to-do-with-geoplot still a WIP).
