### 2018-02-02 ResidentMario/missingno#49 @ResidentMario
Done in 51b67d. The Python versions badge isn't working, so I've just added that to the textual description.
