### 2018-02-02 ResidentMario/missingno#49 @rhiever
Does this badge work for you?
