### 2018-02-02 ResidentMario/missingno#49 @ResidentMario
Ah, right, I forgot you could specify arbitrary characters that way. I went ahead and added a `3.4+` badge, thanks.
