### 2018-02-02 0741c0a
Docstrings, start tests.
