+++ missingno/missingno.py
@@ -12,40 +12,27 @@ from .utils import nullity_filter, nullity_sort
 def matrix(df,
            filter=None, n=0, p=0, sort=None,
            figsize=(25, 10), width_ratios=(15, 1), color=(0.25, 0.25, 0.25),
-           fontsize=16, labels=None, sparkline=True, inline=True,
+           fontsize=16, labels=None, sparkline=True, inline=False,
            freq=None):
     """
-    Presents a `matplotlib` matrix visualization of the nullity of the given DataFrame.
+    A matrix visualization of the nullity of the given DataFrame.
 
-    Note that for the default `figsize` 250 is a soft display limit: specifying a number of records greater than
-    approximately this value will cause certain records to show up in the sparkline but not in the matrix, which can
-    be confusing.
+    For optimal performance, please stay within 250 rows and 50 columns.
 
-
-    The default vertical display will fit up to 50 columns. If more than 50 columns are specified and the labels
-    parameter is left unspecified the visualization will automatically drop the labels as they will not be very
-    readable. You can override this behavior using `labels=True` and your own `fontsize` parameter.
-
-    :param df: The DataFrame whose completeness is being nullity matrix mapped.
-    :param filter: The filter to apply to the heatmap. Should be one of "top", "bottom", or None (default). See
-    `nullity_filter()` for more information.
-    :param n: The cap on the number of columns to include in the filtered DataFrame. See  `nullity_filter()` for
-    more information.
-    :param p: The cap on the percentage fill of the columns in the filtered DataFrame. See  `nullity_filter()` for
-    more information.
-    :param sort: The sort to apply to the heatmap. Should be one of "ascending", "descending", or None. See
-    `nullity_sort()` for more information.
-    :param figsize: The size of the figure to display. This is a `matplotlib` parameter.
-    For the vertical configuration this defaults to (20, 10); the horizontal configuration computes a sliding value
-    by default based on the number of columns that need to be displayed.
-    :param fontsize: The figure's font size. This default to 16.
-    :param labels: Whether or not to display the column names. Would need to be turned off on particularly large
-    displays. Defaults to True.
+    :param df: The `DataFrame` being mapped.
+    :param filter: The filter to apply to the heatmap. Should be one of "top", "bottom", or None (default).
+    :param n: The max number of columns to include in the filtered DataFrame.
+    :param p: The max percentage fill of the columns in the filtered DataFrame.
+    :param sort: The sort to apply to the heatmap. Should be one of "ascending", "descending", or None (default).
+    :param figsize: The size of the figure to display.
+    :param fontsize: The figure's font size. Default to 16.
+    :param labels: If specified, what labels to sue for the column names. Defaults to the underlying data labels when
+    there are 50 columns or less, and no labels when there are more than 50 columns.
     :param sparkline: Whether or not to display the sparkline. Defaults to True.
-    :param width_ratios: The ratio of the width of the matrix to the width of the sparkline. Defaults to `(15,
-    1)`. Does nothing if `sparkline=False`.
-    :param color: The color of the filled columns. Default is a medium dark gray: the RGB multiple `(0.25, 0.25, 0.25)`.
-    :return: If `inline` is True, the underlying `matplotlib.figure` object. Else, nothing.
+    :param width_ratios: The ratio of the width of the matrix to the width of the sparkline. Defaults to `(15, 1)`.
+    Does nothing if `sparkline=False`.
[…]
