### 2018-02-02 3db1470
Add testing instructions to CONTRIBUTING.md
