### 2018-02-03 482bea8
Finish tests, corrections to core functions.
