+++ quilt.yml
@@ -1,0 +1,2 @@
+packages:
+  - ResidentMario/missingno_data
