+++ README.md
@@ -18,6 +18,7 @@ Then to load the data into memory:
 
     >>> from quilt.data.ResidentMario import missingno_data
     >>> collisions = missingno_data.nyc_injurious_collisions()
+    >>> collisions = collisions.replace("nan", np.nan)
 
 The rest of this walkthrough will draw from this `collisions` dataset. I additionally define **nullity** to mean 
 whether a particular variable is filled in or not.
