### 2018-02-03 7832e2d
Move configuration details to a separate file.
