+++ README.md
@@ -190,118 +190,12 @@ projection&mdash;that is, none at all. Not pretty, but functional.
 * `geoplot` requires the [`shapely`](https://github.com/Toblerity/Shapely) and [`descartes`](https://pypi.python.org/pypi/descartes) libraries, which are
 ancillary to the rest of this package and are thus optional dependencies.
 
-That concludes our tour of `missingno`!
+That concludes our tour of `missingno`.
 
-For further details take a look at [this blog post](http://www.residentmar.io/2016/06/12/null-and-missing-data-python.html).
-
-## Other
-
-### Sorting and filtering
-
-`missingno` also provides utility functions for filtering records in your dataset based on completion. These are
-useful in particular for filtering through and drilling down into particularly large datasets whose data nullity
-issues might otherwise be very hard to visualize or understand.
-
-Let's first apply a `nullity_filter()` to the data. The `filter` parameter controls which result set we
-want: either `filter=top` or `filter=bottom`. The `n` parameter controls the maximum number of columns that you want:
- so for example `n=5` makes sure we get *at most* five results. Finally, `p` controls the percentage cutoff. If
- `filter=bottom`, then `p=0.9`  makes sure that our columns are *at most*  90% complete; if `filter=top` we get
- columns which are *at least* 90% complete.
-
-For example, the following query filtered down to only at most 15 columns which are not completely filled.
-
-    >>> filtered_data = msno.nullity_filter(data, filter='bottom', n=15, p=0.999) # or filter='top'
-    >>> msno.matrix(filtered_data.sample(250))
-
-![alt text][matrix_filtered]
-
-[matrix_filtered]: http://i.imgur.com/UF6hmL8.png
-
-`nullity_sort()` simply reshuffles your rows by completeness, in either `ascending` or `descending` order. Since it
-doesn't affect the underlying data it's mainly useful for `matrix` visualization:
-
-
-    >>> sorted_data = msno.nullity_sort(data, sort='descending') # or sort='ascending'
-    >>> msno.matrix(sorted_data.sample(250))
-
-![alt text][matrix_sorted]
-
-[matrix_sorted]: http://i.imgur.com/qL6zNQj.png
-
-These methods work inline within the visualization methods themselves. For instance, the following is perfectly valid:
-
-    >>> msno.matrix(data.sample(250), filter='top', n=5, p=0.9, sort='ascending')
-
-### Visual configuration
-#### Lesser parameters
-
-Each of the visualizations provides a further set of lesser configuration parameters for visually tweaking the display.
-
-`matrix`, `bar`, `heatmap`, `dendrogram`, and `geoplot` all provide:
-
-* `figsize`: The size of the figure to display. This is a `matplotlib` parameter which defaults to `(20, 12)`, except
- for large `dendrogram` visualizations, which compute a height on the fly based on the number of variables to display.
-* `fontsize`: The figure's font size. The default is `16`.
-* `labels`: Whether or not to display the column names. For `matrix` this defaults to `True` for `<=50` variables and
- `False` for `>50`. It always defaults to `True` for `dendrogram` and `heatmap`.
-* `inline`: Defaults to `True`, in which case the chart is plotted and nothing is returned. If this is set to `False`
-the methods omit plotting and return their visualizations instead.
-
-`matrix` also provides:
-* `sparkline`: Set this to `False` to not draw the sparkline.
-* `freq`: If you are working with timeseries data (a `pandas` `DataFrame` with a `PeriodIndex` or `DatetimeIndex`)
-you can specify and display a [choice of offset](http://pandas.pydata.org/pandas-docs/stable/timeseries.html#timeseries-offset-aliases).
[…]
