### 2018-02-03 ResidentMario/missingno#54 @ResidentMario
I've split out the advanced configuration details in the `README.md` into a separate linked-to `CONFIGURATION.md` file in the repository, keeping the focus on the quickstart as much as possible. I won't do it right now, but yeah, maybe I'll set up a website for this repo eventually. I'll leave this issue open for nowc as a prompt to do so.
