### 2018-02-03 ResidentMario/missingno#53 @ResidentMario
Done, now returning the most significant axis object.
