### 2018-02-03 ResidentMario/missingno#50 @ResidentMario
This bothered me also. I've fixed this an alternative way however, by moving the y text labels from the top position to the bottom position. I think this is a much more elegant solution to this problem. See the `README.md` for the current "look".
