### 2018-02-03 ResidentMario/missingno#45 @ResidentMario
Done. A sample collisions dataset is now externalized via `quilt` (you can also [just download it directly from a companion rep](https://github.com/ResidentMario/missingno-data)), and I've updated all of the Quickstart visualizations to use it. I've removed the housing dataset, which I no longer have at hand, would be hard to recreate, and mostly just pads the length out IMHO.
