### 2018-02-04 c6cedbd
Rewrite geoplot to output to aggplot.
