+++ missingno/missingno.py
@@ -5,7 +5,6 @@ import matplotlib.pyplot as plt
 from scipy.cluster import hierarchy
 import seaborn as sns
 import pandas as pd
-from mpl_toolkits.axes_grid1 import make_axes_locatable
 from .utils import nullity_filter, nullity_sort
 
 
@@ -443,222 +442,42 @@ def _calculate_geographic_nullity(geo_group, x_col, y_col):
         return points, np.nan
 
 
-def geoplot(df, x=None, y=None, coordinates=None, by=None, geometry=None, cutoff=None, histogram=False,
-            figsize=(25, 10), fontsize=8, inline=False):
+def geoplot(df,
+            filter=None, n=0, p=0, sort=None,
+            x=None, y=None, coordinates=None, figsize=(25, 10), inline=False,
+            by=None, cmap='Reds', vmin=0, vmax=1, **kwargs):
     """
     Generates a geographical data nullity heatmap, which shows the distribution of missing data across geographic
-    regions. The precise output depends on the inputs provided. In increasing order of usefulness:
-
-    *   If no geographical context is provided, a quadtree is computed and nullities are rendered as abstract
-        geopgrahical squares.
-    *   If geographical context is provided in the form of a column of geographies (region, borough. ZIP code,
-        etc.) in the `DataFrame`, convex hulls are computed for each of the point groups and the heatmap is generated
-        within them.
-    *   If geographical context is provided *and* a separate geometry is provided, a heatmap is generated for each
-        point group within this geograpby instead.
-
-    :param df: The DataFrame whose completeness is being mapped.
-    :param x: The x variable: probably a coordinate (longitude), possibly some other floating point value. May be a
-    string (pointing to a column of df) or an iterable.
-    :param y: The y variable: probably a coordinate (latitude), possibly some other floating point value. May be a
-    string (pointing to a column of df) or an iterable.
-    :param coordinates: A coordinate tuple iterable, or column thereof in the given DataFrame. One of x AND y OR
-    coordinates must be specified, but not both.
-    :param by: If you would like to aggregate your geometry by some geospatial attribute of the underlying DataFrame,
-    name that column here.
-    :param geometry: If you would like to provide your own geometries for your aggregation, instead of relying on
-    (functional, but not pretty) convex hulls, provide them here. This parameter is expected to be a dict or Series
-    of `shapely.Polygon` or `shapely.MultiPolygon` objects. It's ignored if `by` is not specified.
-    :param cutoff: If no aggregation is specified, this parameter sets the minimum number of observations to include in
-    each square. If not provided, set to 50 or 5% of the total size of the dataset, whichever is smaller. If `by` is
-    specified this parameter is ignored.
-    :param figsize: The size of the figure to display. This is a `matplotlib` parameter which defaults to (25, 10).
-    :param histogram: Whether or not to plot a histogram of data distributions below the map. Defaults to False.
-    :param fontsize: If `hist` is specified, this parameter specifies the size of the tick labels. Ignored if `hist`
-    is not specified. Defaults to 8.
-    :param inline: Whether or not the figure is inline. If it's not then instead of getting plotted, this method will
-    return its figure.
-    :return: If `inline` is False, the underlying `matplotlib.figure` object. Else, nothing.
+    regions. The precise output depends on the inputs provided. If no geographical context is provided, a quadtree
[…]
