+++ tests/viz_tests.py
@@ -122,3 +122,25 @@ class TestDendrogram(unittest.TestCase):
     def test_method_dendrogram(self):
         msno.dendrogram(self.simple_df, method='single')
         return plt.gcf()
+
+
+class TestGeoplot(unittest.TestCase):
+    """Smoke tests only. The main function operations are handled by and tested in the `geoplot` package."""
+    # TODO: Add more tests.
+
+    def setUp(self):
+        np.random.seed(42)
+        simple_df = pd.DataFrame((np.random.random((20, 10))), columns=range(0, 10))
+        simple_df = simple_df.add_prefix("r")
+        self.x_y_df = simple_df
+        self.coord_df = simple_df.assign(coords=simple_df.apply(lambda srs: (srs['r0'], srs['r1']), axis='columns'))
+
+    @pytest.mark.mpl_image_compare
+    def test_x_y_geoplot(self):
+        msno.geoplot(self.x_y_df, x='r0', y='r1')
+        return plt.gcf()
+
+    @pytest.mark.mpl_image_compare
+    def test_coordinates_geoplot(self):
+        msno.geoplot(self.coord_df, coordinates='coords')
+        return plt.gcf()
