### 2018-02-04 6c1d5a6
Finish geoplot rewrite.
