+++ tests/viz_tests.py
@@ -125,7 +125,7 @@ class TestDendrogram(unittest.TestCase):
 
 
 class TestGeoplot(unittest.TestCase):
-    """Smoke tests only. The main function operations are handled by and tested in the `geoplot` package."""
+    """Integration tests only. The main function operations are handled by and tested in the `geoplot` package."""
     # TODO: Add more tests.
 
     def setUp(self):
@@ -133,14 +133,8 @@ class TestGeoplot(unittest.TestCase):
         simple_df = pd.DataFrame((np.random.random((20, 10))), columns=range(0, 10))
         simple_df = simple_df.add_prefix("r")
         self.x_y_df = simple_df
-        self.coord_df = simple_df.assign(coords=simple_df.apply(lambda srs: (srs['r0'], srs['r1']), axis='columns'))
 
     @pytest.mark.mpl_image_compare
-    def test_x_y_geoplot(self):
+    def test_geoplot_quadtree(self):
         msno.geoplot(self.x_y_df, x='r0', y='r1')
-        return plt.gcf()
-
-    @pytest.mark.mpl_image_compare
-    def test_coordinates_geoplot(self):
-        msno.geoplot(self.coord_df, coordinates='coords')
         return plt.gcf()
