### 2018-02-04 ResidentMario/missingno#51 @ResidentMario
I've been doing pretty extensive work on this library over the last few days, and I haven't run into this issue. It looks to me that the _masking_ is correct, but the colors are off: `0.2` corresponds with the most neutral color here, while _actual_ neutral values at 0.0 are assigned a darker red.

Do you still get this issue if you follow the new (significantly simpler data-wise) Quickstart?

Was this plot generated using `matplotlib 2.0`?
