### 2018-02-04 ResidentMario/missingno#47 @ResidentMario
Done, now living at `viz_tests.py`.
