### 2018-02-04 a894084
Remove leftover method.
