### 2018-02-04 ResidentMario/missingno#46 @ResidentMario
Done. The `_calculate_geographic_nullity()` comment is no longer relevant because that methodology has been moved out of the package entirely (see discussion elsewhere).
