### 2018-02-04 ResidentMario/missingno#59 @ResidentMario
It's quite possible that this is because `geoplot` is simply broken. It was never very well utilized, and hence, not terrifically well-maintained.

The timeline here is that after implementing this method experimentally in this package, I went and implemented a more general method, `aggplot`, in a package I published last January called [`geoplot`](https://github.com/ResidentMario/geoplot). That method is vastly better tested, and I believe also has significantly better performance characteristics. I also have not personally seen any uses of this geospatial plotting method "in the wild".

After thinking about this some, I decided to throw out the current `geoplot` method body and replace it with a thin interface to `geoplot.aggplot`. The existing methodology is a maintenance burden, `aggplot` provides a better implementation, and I doubt this will cause much user disruption. It's a backwards-incompatible change, but all things considered I'm comfortable making it.

The new method is now implemented. Let me know if you have any issues with it.
