### 2018-02-04 ResidentMario/missingno#60 @ResidentMario
This was because `geoplot` requires some geospatial libraries which can be very difficult to install. Additionally, it was far and away the least used method in the library (I've never seen it used in the wild). It didn't seem justifiable to me that I add to the prerequisite burden for it, so I made the required libraries optional dependencies (in the style of, say, `xlrd` for `pandas`).

The rewritten `geoplot` method (see [here](https://github.com/ResidentMario/missingno/issues/59#issuecomment-362872126) for details on why I rewrote it) relies on the `geoplot` and `geopandas` packages. Same deal holds, though. I've gone ahead and marked them clearly as optional dependencies in the `README.md`.
