### 2018-02-04 #547 @ResidentMario
@zkamvar @rhiever I've addressed the bulk of your code comments. You should take a look at the library again (open new issues etcetera), as it's changed significantly in the process.

My current plan is to finish up the library work that was uncovered in the review process and publish a `0.4.0` version with the changes. I'll finish up the remaining comments on the paper last.
