### 2018-02-04 3ea54ea
Reverse the colormap.
