### 2018-02-05 ResidentMario/missingno#51 @rhiever
```Python
In [1]: from matplotlib import __version__

In [2]: __version__
Out[2]: '2.1.2'
```

Looks like I'm reproducing the heatmap in the README fine now. Good work with the revised quickstart. One small request: Please take the `>>>` out of the code examples; I end up having to remove those in my notebook anyway.
