### 2018-02-05 #547 @Kevin-Mattheus-Moerman
@ResidentMario about:
> My current plan is to finish up the library work that was uncovered in the review process and publish a 0.4.0 version with the changes.

Sounds good. I recommend minting this new version once we are good to accept. That way the paper/DOI will also point to your latest version and it wont be behind on any final edits.
