### 2018-02-05 #547 @Kevin-Mattheus-Moerman
@zkamvar @rhiever has @ResidentMario addressed your comments sufficiently?
