### 2018-02-05 #547 @rhiever
I've checked off the items on my list that have been addressed.
