### 2018-02-06 b60918b
Update paper.
