+++ paper.md
@@ -10,7 +10,7 @@ authors:
 affiliations:
  - name: Independent
    index: 1
-date: 28 January 2018
+date: 6 Febuary 2018
 bibliography: paper.bib
 ---
 
@@ -27,23 +27,26 @@ masks based on whether individual entries contain data (which evaluates to true)
 false). This "nullity matrix" is then exposed to user assessment through a variety of special-purpose data 
 visualizations.
 
-The simplest tools, the bar chart and matrix display, are literal translations of a data table's 
-nullity matrix, and are effective for snapshotting general patterns.
+The simplest tool, the bar chart, is a snapshot of column-level information:
 
-![](http://i.imgur.com/DdepYwr.png)
-![](http://i.imgur.com/lOTN3tm.png)
+![](https://i.imgur.com/2BxEfOr.png)
+
+The matrix display provides a literal translations of a data table's 
+nullity matrix. It is useful for snapshotting general patterns:
+
+![](https://i.imgur.com/gWuXKEr.png)
 
 A heatmap provides a methodology for examining relationships within pairs of variables.
 
-![](http://i.imgur.com/ESsZRlY.png)
+![](https://i.imgur.com/JalSKyE.png)
 
-Higher-cardinality data nullity correlations can be understood using a hierarchically clustered dendrogram.
+Higher-cardinality data nullity correlations can be understood using a hierarchically clustered dendrogram:
 
-![](http://i.imgur.com/6ZBC4af.png)
+![](https://i.imgur.com/oIiR4ct.png)
 
-Finally, geospatial data dependencies are viewable using an approach based on the quadtree or convex hull algorithm.
+Finally, geospatial data dependencies are viewable using an approach based on the quadtree or convex hull algorithm:
 
-![large-geoplot](http://i.imgur.com/4dtGhig.png)
+![](https://i.imgur.com/0aaNa9Q.png)
 
 The visualizations are consciously designed to be as effective as possible
 at uncovering missing data patterns both between and within columns of data, and hence, to help its users build more 
