+++ paper.bib
@@ -39,3 +39,12 @@
     year=2010,
     pages=51-56
 }
+
+@misc{seaborn,
+    author={Michael Waskom and others},
+    title        = {mwaskom/seaborn: v0.8.1 (September 2017)},
+    month        = sep,
+    year         = 2017,
+    doi          = {10.5281/zenodo.883859},
+    url          = {https://doi.org/10.5281/zenodo.883859}
+}
