### 2018-02-06 ResidentMario/missingno#44 @ResidentMario
Think I have this covered now.
