+++ paper.bib
@@ -11,7 +11,7 @@
     title =     {{SciPy}: Open source scientific tools for {Python}},
     year =      {2001--},
     url = "http://www.scipy.org/",
-    note = {[Online; accessed <today>]}
+    note = {[Online; accessed <today>]},
 }
 
 @article{numpy,
@@ -20,7 +20,8 @@
     journal={Computing in Science & Engineering},
     volume=13,
     pages=22-30,
-    year=2011
+    year=2011,
+    doi={10.1109/MCSE.2011.37}
 }
 
 @article{matplotlib,
@@ -29,7 +30,8 @@
     year=2007,
     volume=9,
     pages=90-95,
-    journal={Computing in Science & Engineering}
+    journal={Computing in Science & Engineering},
+    doi={10.1109/MCSE.2007.55}
 }
 
 @article{pandas,
