### 2018-02-06 ResidentMario/missingno#48 @ResidentMario
Done: https://github.com/ResidentMario/missingno/tree/dev
