### 2018-02-06 ResidentMario/missingno#43 @ResidentMario
@Kevin-Mattheus-Moerman  I've put in figures for each of the functions. Hopefully that doesn't make it overlong?
