### 2018-02-06 #547 @ResidentMario
I believe the other remaining non-paper issues have been addressed. I've finished tweaking the paper to my understanding of it, but:

* It's been a long time since I've used BibTeX so I'm not sure I've got the citations formatted right. Is it SOP that they aren't showing up in the pre-render `whedon` is generating?
* Maybe including examples of all of the plots is too much for the paper, lengthwise. missingno#43
