### 2018-02-07 ResidentMario/missingno#43 @Kevin-Mattheus-Moerman
The length is fine. Some of the typos are still present in the paper. To fix the references use `[@articleName]` to cite items in the paper. This will make them appear in the text and the reference list.
