### 2018-02-07 #547 @Kevin-Mattheus-Moerman
@ResidentMario your .bib file looks good. It just looks like you are not citing the articles in the paper. You cite them like this for instance:
`[@pandas]`
