### 2018-02-07 ResidentMario/missingno#43 @ResidentMario
They are? Damn, OK, thanks for the thoroughness, I'll go over it again. Maybe I've edited in new ones!
