### 2018-02-07 ResidentMario/missingno#43 @Kevin-Mattheus-Moerman
See tick boxes above. Its the "easily to.. understand". Should probably be "easy to.... understand".
