### 2018-02-13 6c7e307
Add environment.yml to setup binder using conda
