+++ environment.yml
@@ -1,0 +1,7 @@
+channels:
+  - conda-forge
+  - defaults
+dependencies:
+  - geoplot
+  - pip:
+      - quilt
