### 2018-02-13 3eb8341
Corrections and citations to paper.
