### 2018-02-13 #547 @rhiever
Meant to note here that all of my blockers have been resolved, so this paper has my 👍.
