### 2018-02-13 74f1e47
Set flags for 0.4.0
