### 2018-02-13 #547 @ResidentMario
@Kevin-Mattheus-Moerman I've made the corrections to the paper in [missingno#43](https://github.com/ResidentMario/missingno/issues/43). I've also pushed the new version, 0.4.0 to PyPi. The version under review should be updated to 0.4.0, but I'm not sure how to do that. Let me know if there's anything else that needs doing (cc @zkamvar: the paper references have been resolved).
