### 2018-02-13 #547 @zkamvar
@ResidentMario, my blockers have been resolved 👍

The only recommendation I have left would be to name the packages explicitly in text with the citation, for which I've created a PR.
