### 2018-02-14 #547 @Kevin-Mattheus-Moerman
@arfon can you help with this:

> The version under review should be updated to 0.4.0, but I'm not sure how to do that.
