### 2018-02-14 #547 @Kevin-Mattheus-Moerman
@ResidentMario can you work on @zkamvar 's [PR](https://github.com/ResidentMario/missingno/pull/64), seems like a good suggestion.
