### 2018-02-19 #547 @ResidentMario
> the DOI of the reviewed software available

Apologies---what do you mean by this? A little unclear to me. :)
