### 2018-02-19 #547 @Kevin-Mattheus-Moerman
Basically you have to archive your work on a repository with a DOI.

See here: http://joss.theoj.org/about#author_guidelines
> Upon successful completion of the review, deposit a copy of your (updated) repository with a data-archiving service such as Zenodo or figshare, issue a DOI for the archive, and update the review issue thread with your DOI.

I use Zenodo (but other services like figshare can be used as well) and these steps: https://guides.github.com/activities/citable-code/

Once you have a DOI link paste it here. Let me know if you need help.
