### 2018-02-25 #547 @Kevin-Mattheus-Moerman
@ResidentMario how are you doing? Have you been able to archive the reviewed software in a service like [Zenodo](https://zenodo.org/) (again these steps are handy: https://guides.github.com/activities/citable-code/). One you provide the archived version DOI we can move on to the next steps.
