### 2018-02-25 #547 @Kevin-Mattheus-Moerman
@rhiever @zkamvar to wrap things up would you be able to tick the remaining boxes at the top of this issue?
