### 2018-02-25 #547 @rhiever
Two of my non-blockers have yet to be addressed (still unchecked), but I approve the paper/package regardless.
